+++ src/main/java/org/synyx/urlaubsverwaltung/absence/web/AbsenceOverviewMonthPersonDto.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 public class AbsenceOverviewMonthPersonDto {
 
-    private final Integer id;
+    private final Long id;
     private final String firstName;
     private final String lastName;
     private final String gravatarUrl;
     private final List<AbsenceOverviewPersonDayDto> days;
 
-    AbsenceOverviewMonthPersonDto(Integer id, String firstName, String lastName, String gravatarUrl, List<AbsenceOverviewPersonDayDto> days) {
+    AbsenceOverviewMonthPersonDto(Long id, String firstName, String lastName, String gravatarUrl, List<AbsenceOverviewPersonDayDto> days) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -18,7 +18,7 @@ public class AbsenceOverviewMonthPersonDto {
         this.days = days;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
