+++ src/main/java/org/synyx/urlaubsverwaltung/person/PersonCreatedEvent.java
@@ -5,13 +5,13 @@ import org.springframework.lang.Nullable;
 
 public class PersonCreatedEvent extends ApplicationEvent {
 
-    private final Integer personId;
+    private final Long personId;
     private final String personNiceName;
     private final String username;
     private final String email;
     private final boolean active;
 
-    PersonCreatedEvent(Object source, Integer personId, String personNiceName, String username, @Nullable String email, boolean active) {
+    PersonCreatedEvent(Object source, Long personId, String personNiceName, String username, @Nullable String email, boolean active) {
         super(source);
         this.personId = personId;
         this.personNiceName = personNiceName;
@@ -20,7 +20,7 @@ public class PersonCreatedEvent extends ApplicationEvent {
         this.active = active;
     }
 
-    Integer getPersonId() {
+    Long getPersonId() {
         return personId;
     }
