+++ src/main/java/org/synyx/urlaubsverwaltung/person/web/PersonDetailsViewController.java
@@ -58,7 +58,7 @@ public class PersonDetailsViewController {
     }
 
     @GetMapping("/person/{personId}")
-    public String showPersonInformation(@PathVariable("personId") Integer personId,
+    public String showPersonInformation(@PathVariable("personId") Long personId,
                                         @RequestParam(value = "year", required = false) Optional<Integer> requestedYear,
                                         Model model) throws UnknownPersonException {
