+++ src/main/java/org/synyx/urlaubsverwaltung/person/web/PersonsViewController.java
@@ -82,7 +82,7 @@ public class PersonsViewController {
     @PreAuthorize(IS_PRIVILEGED_USER)
     @GetMapping("/person")
     public String showPerson(@RequestParam(value = "active", required = false, defaultValue = "true") boolean active,
-                             @RequestParam(value = "department", required = false) Optional<Integer> requestedDepartmentId,
+                             @RequestParam(value = "department", required = false) Optional<Long> requestedDepartmentId,
                              @RequestParam(value = "year", required = false) Optional<Integer> requestedYear,
                              @RequestParam(value = "query", required = false, defaultValue = "") String query,
                              @SortDefault.SortDefaults({
@@ -115,7 +115,7 @@ public class PersonsViewController {
         Page<Person> personPage = null;
 
         if (requestedDepartmentId.isPresent()) {
-            final Integer departmentId = requestedDepartmentId.get();
+            final Long departmentId = requestedDepartmentId.get();
             final Department department = departmentService.getDepartmentById(departmentId)
                 .orElseThrow(() -> new UnknownDepartmentException(departmentId));
